Add deleteitem permissions

+++ backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, niceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -42,9 +43,14 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title}`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
     // 2. check if they own the item, or have permissions
-    // TODO:
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.requrest.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+
+    if(!ownsItem || hasPermissions){
+      throw new Error('You do not have permission to do that1');
+    }
     // 3. delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -165,6 +171,31 @@ const Mutations = {
     // 8. return new user
     return updatedUser;
   },
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. check if they are logged in
+    if(!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+    // 2. query the current user
+    const currentUser = await ctx.db.query.user({
+      where: {
+        id: ctx.request.userId,
+      },
+    },info);
+    // 3. check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // 4. update the permissions
+    return ctx.db.mutation.updateUser({
+      data: {
+        permissions: {
+          set: args.permissions,
+        },
+      },
+      where: {
+        id: args.userId
+      }
+    }, info);
+  };
 };
 
 module.exports = Mutations;
